add resolvers and typeDefs files for GraphQL schema
