+++ package.json
@@ -16,7 +16,8 @@
   "dependencies": {
     "@apollo/server": "4.10.5",
     "@prisma/client": "5.18.0",
-    "graphql": "16.9.0"
+    "graphql": "16.9.0",
+    "graphql-tag": "2.12.6"
   },
   "devDependencies": {
     "@eslint/js": "9.8.0",
