+++ src/index.ts
@@ -1,17 +1,7 @@
 import { ApolloServer } from '@apollo/server';
 import { startStandaloneServer } from '@apollo/server/standalone';
-
-const typeDefs = `#graphql
-  type Query {
-    hello: String!
-  }
-`;
-
-const resolvers = {
-  Query: {
-    hello: () => 'Hello, World!',
-  },
-};
+import typeDefs from './graphQL/typeDefs.js';
+import resolvers from './graphQL/resolvers.js';
 
 const server = new ApolloServer({
   typeDefs,
